pyproject.toml
[project]
name = "timebot"
version = "0.1.0"
description = "A Discord bot that shows the current time in relation to a specified timezone."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asqlite>=2.0.0",
    "discord-py>=2.6.4",
    "jishaku>=2.6.3",
    "toml>=0.10.2",
]

[tool.black]
line-length = 125

[tool.isort]
profile = "black"
line_length = 125
force_alphabetical_sort_within_sections = true
combine_as_imports = true

[tool.pyright]
typeCheckingMode = "standard"
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportFunctionMemberAccess = false

[tool.ruff]
exclude = [".venv"]
line-length = 125
target-version = "py312"

[tool.ruff.lint]
select = ["ASYNC", "C4", "E", "F", "FA", "ISC", "N", "PERF", "PIE", "PL", "RET", "RUF", "SIM", "T20", "TID", "TRY"]
ignore = [
    "F403",
    "F405",
    "N804",  # pylance already does it
    "PLC0414",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "SIM105",
    "TRY003",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"